pyproject.toml
[tool.poetry]
name = "css-chat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
langchain = "^0.3.7"
openai = "^1.54.3"
langchain-community = "^0.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
